backend/requirements.txt
asgiref>=3.8.1,<4.0
celery>=5.4,<6.0
channels>=4.1,<5.0
channels-redis>=4.2,<5.0
click>=8.1,<9.0
django>=5.0,<6.0
django-celery-beat>=2.6,<3.0
django-cors-headers>=4.3,<5.0
django-environ>=0.11,<0.12
django-filter>=24.1,<25.0
django-storages>=1.14,<2.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
drf-spectacular>=0.27,<0.28
gunicorn>=21.2,<22.0
numpy>=1.26,<2.0
packaging>=24.0,<25.0
pandas>=2.2,<3.0
Pillow>=10.2,<11.0
psycopg[binary]>=3.1,<4.0
python-dateutil>=2.8,<3.0
python-dotenv>=1.0,<2.0
redis>=5.0,<6.0
requests>=2.31,<3.0
scikit-learn>=1.4,<2.0
spacy>=3.7,<4.0
tensorflow>=2.14,<3.0
torch>=2.2,<3.0
transformers>=4.38,<5.0
uvicorn[standard]>=0.27,<0.28
Django==5.0.4
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.9.0
django-filter==25.1
channels==4.3.1
channels-redis==4.1.0
celery==5.3.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
psycopg2-binary==2.9.11
redis==7.0.1
python-dotenv==1.0.1
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.4.1.post1
spacy==3.7.2
transformers==4.38.1
torch==2.2.1
tensorflow==2.15.0
nltk==3.8.1
shap==0.43.0
boto3==1.34.23
drf-spectacular==0.29.0
whitenoise==6.11.0
daphne==4.2.1
